Switch from Java 8 to Java 11
When this is applied you need to change the JDK you use for development
from Java 8 to at least Java 11.
