+++ .github/workflows/bygg_alle_brancher.yml
@@ -17,6 +17,12 @@ jobs:
       - name: 'Pull repo'
         uses: actions/checkout@v1
 
+      # JAVA 11
+      - name: 'Java 11'
+        uses: actions/setup-java@v1.3.0
+        with:
+          java-version: 11
+
       # BYGGER DOCKER CONTAINER
       - name: 'Bygg'
         env:
