+++ .github/workflows/bygg_og_deploy_q1branch.yml
@@ -26,6 +26,12 @@ jobs:
       - name: 'Setter Image'
         run: echo "::set-env name=IMAGE::docker.pkg.github.com/${{ github.repository }}/eessi-pensjon-saksbehandling-api:${{ env.DATE }}---${{ env.COMMIT_HASH }}"
 
+      # JAVA 11
+      - name: 'Java 11'
+        uses: actions/setup-java@v1.3.0
+        with:
+          java-version: 11
+
       # BYGGER DOCKER CONTAINER
       - name: 'Bygg og publiser docker image'
         env:
