+++ build.gradle
@@ -30,6 +30,8 @@ plugins {
     id 'com.adarshr.test-logger' version '2.0.0'
 }
 
+assert JavaVersion.current().isJava11Compatible(): "Java 11 or newer is required"
+
 apply plugin: 'kotlin'
 apply plugin: 'kotlin-spring'
 apply plugin: 'org.springframework.boot'
@@ -52,18 +54,22 @@ test {
 
 group = 'no.nav.eessi'
 version = '0.0.1'
-sourceCompatibility = 1.8
+
+java {
+    sourceCompatibility = JavaVersion.VERSION_11
+    targetCompatibility = JavaVersion.VERSION_11
+}
 
 compileKotlin {
     kotlinOptions {
         freeCompilerArgs = ["-Xjsr305=strict"]
-        jvmTarget = "1.8"
+        jvmTarget = "11"
     }
 }
 compileTestKotlin {
     kotlinOptions {
         freeCompilerArgs = ["-Xjsr305=strict"]
-        jvmTarget = "1.8"
+        jvmTarget = "11"
     }
 }
 jar {
