+++ src/test/kotlin/no/nav/eessi/pensjon/api/submit/SubmitControllerTest.kt
@@ -1,20 +1,43 @@
 package no.nav.eessi.pensjon.api.submit
 
+import com.fasterxml.jackson.databind.ObjectMapper
 import com.nhaarman.mockitokotlin2.any
 import com.nhaarman.mockitokotlin2.doNothing
 import com.nhaarman.mockitokotlin2.doReturn
 import com.nhaarman.mockitokotlin2.doThrow
 import com.nhaarman.mockitokotlin2.whenever
+import no.nav.eessi.pensjon.services.kafka.KafkaService
+import no.nav.eessi.pensjon.services.pdf.PdfService
+import no.nav.eessi.pensjon.services.pdf.TemplateService
+import no.nav.eessi.pensjon.services.storage.S3StorageBaseTest
 import no.nav.eessi.pensjon.utils.mapAnyToJson
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
 import org.mockito.Mockito
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
+import org.springframework.kafka.core.DefaultKafkaProducerFactory
+import org.springframework.kafka.core.KafkaTemplate
+import org.springframework.web.client.RestTemplate
+import org.apache.kafka.clients.CommonClientConfigs
+import org.apache.kafka.clients.producer.ProducerConfig
+import org.apache.kafka.common.config.SaslConfigs
+import org.apache.kafka.common.serialization.StringSerializer
 
-class SubmitControllerTest : SubmitBaseTest() {
+class SubmitControllerTest : S3StorageBaseTest() {
+
+    var mapper = ObjectMapper()
+
+    lateinit var receiveSubmissionController : ReceiveSubmissionController
+    lateinit var mockFagmodulRestTemplate : RestTemplate
+    lateinit var pdfService: PdfService
+    lateinit var templateService: TemplateService
+    lateinit var kafkaService : KafkaService
+    lateinit var kafkaTemplate : KafkaTemplate<String, String>
+
 
     @AfterEach fun cleanUpTest() {
         Mockito.reset(mockFagmodulRestTemplate)
@@ -168,5 +191,36 @@ class SubmitControllerTest : SubmitBaseTest() {
 
         val generatedResponse = receiveSubmissionController.getSubmissionAsJson(subject)
         assertEquals(expectedResponse, generatedResponse)
+    }
+
+    fun producerFactory(): DefaultKafkaProducerFactory<String, String> {
+        val properties = mapOf<String, Any>(
+                ProducerConfig.BOOTSTRAP_SERVERS_CONFIG to "brokers",
+                ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG to StringSerializer::class.java,
+                ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG to StringSerializer::class.java,
+                ProducerConfig.CLIENT_ID_CONFIG to "appName",
+                CommonClientConfigs.SECURITY_PROTOCOL_CONFIG to "SASL_SSL",
+                SaslConfigs.SASL_MECHANISM to "PLAIN",
+                SaslConfigs.SASL_JAAS_CONFIG to "org.apache.kafka.common.security.plain.PlainLoginModule required username=\"username\" password=\"password\";"
+        )
+        return DefaultKafkaProducerFactory(properties)
+    }
+
+    @BeforeEach
+    fun init() {
+        kafkaTemplate = Mockito.spy(KafkaTemplate<String, String>(producerFactory()))
+        kafkaService = Mockito.spy(KafkaService(kafkaTemplate))
+
+        mockFagmodulRestTemplate = generateMockFagmodulRestTemplate()
+
+        pdfService = Mockito.spy(PdfService())
+        templateService = Mockito.spy(TemplateService())
+
+        receiveSubmissionController = Mockito.spy(ReceiveSubmissionController(
+                kafkaService,
+                s3storageService,
+                generateMockContextHolder(),
+                templateService
+        ))
     }
 }
